requirements.txt
redis==3.5.3
redisgraph==2.4.4
loguru==0.7.2
pyyaml==6.0.1